requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
reportlab==4.2.0
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Migrate==4.0.7
gunicorn==22.0.0
psycopg2-binary==2.9.9
Celery==5.4.0
redis==5.0.4
limiter==3.6.0
requests==2.31.0
whitenoise==6.7.0
stripe==8.1.0
razorpay==1.4.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
email-validator==2.1.1
